Consolidate scarping functions
